renames model payment to paymentinitiation
