+++ src/main/java/com/plaid/client/response/paymentinitiation/PaymentGetResponse.java
@@ -1,6 +1,6 @@
 package com.plaid.client.response.paymentinitiation;
 
-import com.plaid.client.model.payment.Amount;
+import com.plaid.client.model.paymentinitiation.Amount;
 import com.plaid.client.response.BaseResponse;
 
 import java.util.Date;
