+++ src/main/java/com/plaid/client/response/paymentinitiation/RecipientGetResponse.java
@@ -1,6 +1,6 @@
 package com.plaid.client.response.paymentinitiation;
 
-import com.plaid.client.model.payment.Address;
+import com.plaid.client.model.paymentinitiation.Address;
 import com.plaid.client.response.BaseResponse;
 import java.util.List;
