+++ src/main/java/com/plaid/client/response/paymentinitiation/RecipientListResponse.java
@@ -1,7 +1,7 @@
 package com.plaid.client.response.paymentinitiation;
 
-import com.plaid.client.model.payment.Address;
-import com.plaid.client.model.payment.Recipient;
+import com.plaid.client.model.paymentinitiation.Address;
+import com.plaid.client.model.paymentinitiation.Recipient;
 import com.plaid.client.response.BaseResponse;
 
 import java.util.List;
